requirements.txt
asgiref==3.3.1
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
